requirements.txt
aiohttp
sqlalchemy[asyncio]
asyncpg
python-dotenv
alembic
pydantic